hackathon2/requirements.txt
pandas 
fuzzywuzzy 
python-Levenshtein
sentence_transformers
faiss-cpu
